refactor: 접근제어자 변경 및 컨벤션 수정
