+++ src/main/java/domain/lotto/strategy/SequenceStrategy.java
@@ -1,4 +1,4 @@
-package domain.lotto;
+package domain.lotto.strategy;
 
 import java.util.List;
